requirements.txt
flask
pandas
numpy
scikit-learn
tensorflow
python-docx
openpyxl
gunicorn